requirements.txt
# Dependências do JARVIS 3.0
# Core Framework
flask==3.0.0
flask-socketio==5.3.6
flask-cors==4.0.0
python-socketio==5.10.0

# IA e Machine Learning
openai==1.6.1
transformers==4.36.2
torch>=2.0.0
sentence-transformers==2.2.2

# Processamento de Áudio
gtts==2.4.0
pyttsx3==2.90
librosa==0.10.1
soundfile==0.12.1

# Monitoramento de Sistema
psutil==5.9.7
py-cpuinfo==9.0.0
GPUtil==1.4.0

# Interface Web
jinja2==3.1.2
websockets==12.0
eventlet==0.33.3

# Banco de Dados
sqlalchemy==2.0.25

# Utilitários
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
colorama==0.4.6
rich==13.7.0
click==8.1.7

# Visualização
matplotlib==3.8.2
plotly==5.17.0

# Automação
pyautogui==0.9.54
keyboard==0.13.5

# Processamento de Texto
nltk==3.8.1

# Desenvolvimento
pytest==7.4.4
black==23.12.1
flake8==7.0.0
